backend/requirements.txt
# ── Web Framework ────────────────────────────────────────────────────────────
Flask==3.0.0
Flask-Cors==4.0.0
Werkzeug==3.1.5

# ── Database ──────────────────────────────────────────────────────────────────
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
greenlet==3.3.0

# ── ML / Data Science ─────────────────────────────────────────────────────────
xgboost==3.1.3
scikit-learn==1.6.1
numpy==1.26.2
pandas==2.1.4
joblib==1.3.2
scipy==1.16.3

# ── Firebase / FCM ────────────────────────────────────────────────────────────
firebase_admin==7.1.0
google-api-core==2.30.0
google-auth==2.47.0
google-auth-oauthlib==1.2.4
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.78.1
proto-plus==1.27.1
protobuf==6.33.5

# ── Scheduler ─────────────────────────────────────────────────────────────────
APScheduler==3.11.2
schedule==1.2.2

# ── HTTP / Networking ─────────────────────────────────────────────────────────
requests==2.31.0
requests-oauthlib==2.0.0
urllib3==2.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
anyio==4.12.1

# ── Auth / Security ───────────────────────────────────────────────────────────
cryptography==46.0.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
rsa==4.9.1
PyJWT==2.11.0
oauthlib==3.3.1
CacheControl==0.14.4
msgpack==1.1.2

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2025.2
tzdata==2025.3
tzlocal==5.3.1
six==1.17.0
packaging==25.0
click==8.3.1
blinker==1.9.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
colorama==0.4.6
threadpoolctl==3.6.0
typing_extensions==4.15.0
pycparser==3.0
cffi==2.0.0